Extract `readLocalData` and `writeLocalData`
